Reduce stictness level to allow autoreconf usage
By default certain files are expected according to GNUs standards such
as NEWS, ChangeLog and so on.

Since this project isn't really associated with GNU let's reduce the
strictness level to "foreign" so the lack of these files don't cause
errors.

http://www.gnu.org/software/automake/manual/html_node/Strictness.html

This is principly so autoreconf can be used, eschewing the now redundant
autogen.sh/bootstrap scripts projects have used in the past.

Without using foreign the following errors would be generated:

    Makefile.am: error: required file './NEWS' not found
    Makefile.am: error: required file './AUTHORS' not found
    Makefile.am: error: required file './ChangeLog' not found